jacob tried to stuff a value of 1518 into a one byte number which gcc3 happily accepts but (fortunately) gcc 2.95 caught

+++ servlib/conv_layers/EthConvergenceLayer.h
@@ -72,7 +72,7 @@ public:
      * Current version of the protocol.
      */
     static const u_int8_t ETHCL_VERSION = 0x01;
-    static const u_int8_t MAX_ETHER_PACKET = 1518;
+    static const u_int16_t MAX_ETHER_PACKET = 1518;
     static const u_int16_t ETHERTYPE_DTN = 0xd710;
 
     static const u_int8_t ETHCL_BEACON = 0x01;
